refactor(ingest): remove the XML order form entry point
The JSON sidecar path added in 73de9db replaced the FileCatalyst XML order
form. Nothing produces the XML any more, so this removes the door it came
through and the types that only described it. The import flows themselves are
untouched — Masters, RawMaterialForm, Multitrack and MoveUploadedFiles all
stay registered and keep the same params.

Gone:
  - the Asset workflow, AssetParams, and the copy/sanitize helpers that only
    existed to turn an XML fileList into a directory: copyToDir,
    sanitizeOrderForm, sanitizeDuplicatdPath
  - the two ways in: the /mnt/filecatalyst/workflow/xml/* watcher branch and
    the "AssetIngest" manual trigger
  - ingest.FileList, ingest.File, ingest.JobHistoryLog, ingest.JobLog and
    Metadata.XMLName, all of which only had readers inside Asset, plus the
    xml struct tags now that nothing unmarshals into these
  - thirteen JobProperty fields with no reader anywhere — the FileCatalyst
    transfer details (JobID aside: raw_material.go still reports it on
    failure)
  - wfutils.UnmarshalXMLFile, whose only caller was Asset. MarshalXml stays,
    it writes the SMIL in vx_export_vod

What survives of asset_ingest.go is the OrderForm enum, which AssetJSON
derives from the form key, so the file is now order_form.go. Its enum is left
whole: the members beyond VB and LED-Material have no JSON form yet, but they
name flows that still exist.

Tests: the Asset cases go with the workflow, along with the three XML
fixtures. Test_VBBulk_MasterFlow tested Masters through Asset's XML parsing —
it now builds the same ingest.Metadata as a literal and drives Masters
directly, so it keeps testing the flow that is still used. Moved to
masters_test.go accordingly.
