+++ utils/workflows/encoding_test.go
@@ -2,12 +2,10 @@ package wfutils
 
 import (
 	"github.com/bcc-code/bcc-media-flows/activities"
-	"github.com/bcc-code/bcc-media-flows/paths"
 	"github.com/stretchr/testify/suite"
 	"go.temporal.io/sdk/testsuite"
 	"go.temporal.io/sdk/workflow"
 	"testing"
-	"time"
 )
 
 type UnitTestEncoding struct {
@@ -49,11 +47,6 @@ func UnmarshalJSONTest(ctx workflow.Context) (*testStruct, error) {
 	return UnmarshalJson[testStruct](ctx, []byte("{\"name\":\"test\",\"thing\":1}"))
 }
 
-func UnmarshalXMLFileTest(ctx workflow.Context) (*testStruct, error) {
-	ctx = workflow.WithActivityOptions(ctx, workflow.ActivityOptions{StartToCloseTimeout: 10 * time.Second})
-	return UnmarshalXMLFile[testStruct](ctx, paths.MustParse("./testdata/test.xml"))
-}
-
 func (s *UnitTestEncoding) Test_MarshalXML() {
 	s.env.ExecuteWorkflow(MarshalXMLTest)
 
@@ -63,23 +56,6 @@ func (s *UnitTestEncoding) Test_MarshalXML() {
 	s.NotEmpty(t)
 	s.NoError(err)
 	s.Equal("<testStruct>\n\t<name>test</name>\n\t<thing>1</thing>\n</testStruct>", string(t))
-
-	s.True(s.env.IsWorkflowCompleted())
-	s.NoError(s.env.GetWorkflowError())
-}
-
-func (s *UnitTestEncoding) Test_UnmarshalXMLFile() {
-	s.env.ExecuteWorkflow(UnmarshalXMLFileTest)
-
-	var t testStruct
-	err := s.env.GetWorkflowResult(&t)
-
-	s.NotEmpty(t)
-	s.NoError(err)
-	s.Equal(testStruct{
-		Name:  "test",
-		Thing: 1,
-	}, t)
 
 	s.True(s.env.IsWorkflowCompleted())
 	s.NoError(s.env.GetWorkflowError())
