--- utils/workflows/testdata/test.xml
@@ -1,4 +1,0 @@
-<testStruct>
-    <name>test</name>
-    <thing>1</thing>
-</testStruct>
