--- workflows/ingest/testdata/Upload.xml
@@ -1,37 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<Metadata>
-    <jobPropertyList>
-        <jobID>7001</jobID>
-        <userName>testuser</userName>
-        <companyName></companyName>
-        <sourceIP>10.12.135.200</sourceIP>
-        <userEmail>test@example.com</userEmail>
-        <ingestStation>filecatalyst-01/10.12.135.2</ingestStation>
-        <uploadBitRate>16 kbps</uploadBitRate>
-        <uploadTime>00:00:12</uploadTime>
-        <ftpSiteId>dmz02</ftpSiteId>
-        <fileCount>1</fileCount>
-        <orderForm>Upload</orderForm>
-        <submissionDate>Mon Aug 11 09:00:00 CEST 2026</submissionDate>
-        <lastDateChanged>Mon Aug 11 09:00:00 CEST 2026</lastDateChanged>
-        <status>pending</status>
-        <programid>TEMP - Uploadtest</programid>
-        <program_post>01</program_post>
-        <received_filename>UPLOADTEST</received_filename>
-        <Tags></Tags>
-        <PersonsAppearing></PersonsAppearing>
-        <language>MUL</language>
-        <sender_email>test@example.com</sender_email>
-        <direct_playback>No</direct_playback>
-        <asset_type>MAS</asset_type>
-    </jobPropertyList>
-    <fileList>
-        <file>
-            <FileName>upload1.mov</FileName>
-            <isFolder>false</isFolder>
-            <fileSize>136761</fileSize>
-            <filePath>/files/7001</filePath>
-        </file>
-    </fileList>
-    <jobHistoryLog/>
-</Metadata>
