--- workflows/ingest/testdata/processed/.gitignore
@@ -1,1 +1,0 @@
-*
